requirements.txt
Flask==2.2.0
flask-restful==0.3.9
flask-session==0.4.0
flask-socketio==5.2.0
RPi.GPIO==0.7.1
spidev==3.5
gevent==21.12.0
gevent-websocket==0.10.1
luma.led-matrix==1.6.1